+++ libs/co2/data-access/src/lib/co2-forecast/http/ckan-errors-response.ts
