+++ libs/co2/data-access/src/lib/co2-forecast/http/co2-emission-prognosis-record.ts
