+++ libs/co2/data-access/src/lib/co2-forecast/http/create-co2-forecast-sql-query.ts
@@ -1,4 +1,4 @@
-import { DateQuery } from './date-query';
+import { DateQuery } from '../date-query';
 
 export function createCo2ForecastSqlQuery(dateQuery: DateQuery): string {
   return `SELECT "Minutes5UTC" AS "minutes5Utc"
