+++ libs/co2/data-access/src/lib/co2-forecast/http/energi-data-service-endpoint.ts
