+++ libs/co2/data-access/src/lib/co2-forecast/http/trim-sql-parameter.ts
